Update color map

+++ include/util.hpp
@@ -3,6 +3,7 @@
  k-vernooy
 
  Useful functions for color conversions
+ 
  and definitions for the color map.
 ======================================*/
 
@@ -17,9 +18,9 @@ using std::array;
 int rgb_to_color(int& r, int& g, int& b);
 std::array<int, 3> color_to_rgb(int c);
 Uint32 rgb_to_uint(int r, int g, int b);
-array<int, 3> interpolate_hsl(array<int, 3> hsl1, array<int, 3> hsl2, double interpolator);
-array<int, 3> rgb_to_hsl(array<int, 3> rgb);
-array<int, 3> hsl_to_rgb(array<int, 3> hsl);
+array<double, 3> interpolate_hsl(array<double, 3> hsl1, array<double, 3> hsl2, double interpolator);
+array<double, 3> rgb_to_hsl(array<int, 3> rgb);
+array<int, 3> hsl_to_rgb(array<double, 3> hsl);
 
 
 static std::vector<std::vector<int>> colors = {
